+++ spring-cloud-dataflow-autoconfigure/pom.xml
@@ -33,10 +33,6 @@
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-scheduler-spi</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-deployer-resource-maven</artifactId>
 		</dependency>
 		<dependency>
