+++ spring-cloud-dataflow-core/pom.xml
@@ -58,9 +58,5 @@
 			<groupId>org.springframework.hateoas</groupId>
 			<artifactId>spring-hateoas</artifactId>
 		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-scheduler-spi</artifactId>
-		</dependency>
 	</dependencies>
 </project>
