+++ spring-cloud-dataflow-platform-kubernetes/pom.xml
@@ -19,10 +19,6 @@
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-scheduler-kubernetes</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-starter-kubernetes-config</artifactId>
 		</dependency>
 		<dependency>
