+++ spring-cloud-dataflow-rest-client/src/main/java/org/springframework/cloud/dataflow/rest/client/SchedulerOperations.java
@@ -20,7 +20,7 @@ import java.util.List;
 import java.util.Map;
 
 import org.springframework.cloud.dataflow.rest.resource.ScheduleInfoResource;
-import org.springframework.cloud.scheduler.spi.core.ScheduleInfo;
+import org.springframework.cloud.deployer.scheduler.spi.core.ScheduleInfo;
 import org.springframework.hateoas.PagedModel;
 
 /**
