+++ spring-cloud-dataflow-rest-resource/pom.xml
@@ -80,9 +80,5 @@
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-skipper</artifactId>
 		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-scheduler-spi</artifactId>
-		</dependency>
 		</dependencies>
 </project>
